move utils dir: tests -> tests/python
