+++ dipu/tests/python/unittests/test_profiler.py
@@ -4,7 +4,7 @@ import torch_dipu
 import torchvision.models as models
 from torch.profiler import profile, ProfilerActivity
 from torch_dipu.testing._internal.common_utils import TestCase, run_tests
-from tests.utils.local_eviron import local_eviron
+from tests.python.utils.local_eviron import local_eviron
 
 
 class TestProfiler(TestCase):
